Refactoring of functions to make them more granular and moved some to pcap_device where they belong
