+++ src/pcap_listener.c
@@ -24,49 +24,10 @@
 #include "util.h"
 #include "pcap_device.h"
 
-#define BUFFER_TIMEOUT 2000
-
-static pcap_t *PcapOpenLiveImmediate(const char *source, const int snaplen, const int promisc, const int to_ms, char *errbuf);
 static uint8_t CreateAndAddDevice(struct ListenerModule *lm, const char *name);
 static int AutoPrepDevices(struct ListenerModule *lm, const uint8_t includeLo);
 static int PrepDevices(struct ListenerModule *lm);
-static char *AllocAndBuildPcapFilter(const struct Device *device);
 static void PrintDevices(const struct ListenerModule *lm);
-
-/* Heavily inspired by src/lib/libpcap/pcap-bpf.c from OpenBSD's pcap implementation.
- * We must use pcap_create() and pcap_activate() instead of pcap_open_live() because
- * we need to set the immediate mode flag, which can only be done on an unactivated
- * pcap_t.
- *
- * OpenBSD's libpcap implementation require immediate mode and non-blocking socket
- * in order for poll() (and select()/kevent()) to work properly. This approach works
- * for other Unixes as well, so it's no harm in doing it this way. Using immediate mode
- * with a non-blocking fd makes pcap a bit snappier anyway so it's a win-win.
- * See: https://marc.info/?l=openbsd-tech&m=169878430118943&w=2 for more information.
- * */
-static pcap_t *PcapOpenLiveImmediate(const char *source, const int snaplen, const int promisc, const int to_ms, char *errbuf) {
-  pcap_t *p;
-  int status;
-
-  if ((p = pcap_create(source, errbuf)) == NULL)
-    return (NULL);
-  if ((status = pcap_set_snaplen(p, snaplen)) < 0)
-    goto fail;
-  if ((status = pcap_set_promisc(p, promisc)) < 0)
-    goto fail;
-  if ((status = pcap_set_timeout(p, to_ms)) < 0)
-    goto fail;
-  if ((status = pcap_set_immediate_mode(p, 1)) < 0)
-    goto fail;
-
-  if ((status = pcap_activate(p)) < 0)
-    goto fail;
-  return (p);
-fail:
-  SafeStrncpy(errbuf, pcap_geterr(p), PCAP_ERRBUF_SIZE);
-  pcap_close(p);
-  return (NULL);
-}
 
 static uint8_t CreateAndAddDevice(struct ListenerModule *lm, const char *name) {
   struct Device *dev;
@@ -207,112 +168,6 @@ cleanup:
   return status;
 }
 
-static char *AllocAndBuildPcapFilter(const struct Device *device) {
-  int i;
-  int filterLen = 0;
-  char *filter = NULL;
-
-  assert(device != NULL);
-
-  if (device->inet4_addrs_count > 0 || device->inet6_addrs_count > 0) {
-    filter = ReallocAndAppend(filter, &filterLen, "(");
-  }
-
-  for (i = 0; i < device->inet4_addrs_count; i++) {
-    if (i > 0) {
-      filter = ReallocAndAppend(filter, &filterLen, " or ");
-    }
-    filter = ReallocAndAppend(filter, &filterLen, "ip dst host %s", device->inet4_addrs[i]);
-  }
-
-  if (device->inet4_addrs_count > 0 && device->inet6_addrs_count > 0) {
-    filter = ReallocAndAppend(filter, &filterLen, " or ");
-  }
-
-  for (i = 0; i < device->inet6_addrs_count; i++) {
-    if (i > 0) {
-      filter = ReallocAndAppend(filter, &filterLen, " or ");
-    }
-    filter = ReallocAndAppend(filter, &filterLen, "ip6 dst host %s", device->inet6_addrs[i]);
-  }
-
-  if (device->inet4_addrs_count > 0 || device->inet6_addrs_count > 0) {
-    filter = ReallocAndAppend(filter, &filterLen, ")");
-  }
-
-  filter = ReallocAndAppend(filter, &filterLen, " and (");
-
-  if (configData.tcpPortsLength > 0) {
-    if (configData.tcpPortsLength > 0 && configData.udpPortsLength > 0) {
-      filter = ReallocAndAppend(filter, &filterLen, "(");
-    }
-
-    for (i = 0; i < configData.tcpPortsLength; i++) {
-      if (i > 0) {
-        filter = ReallocAndAppend(filter, &filterLen, " or ");
-      }
-
-      if (IsPortSingle(&configData.tcpPorts[i])) {
-        filter = ReallocAndAppend(filter, &filterLen, "tcp dst port %d", configData.tcpPorts[i].single);
-      } else {
-        /* OpenBSD's libpcap doesn't support portrange */
-#ifdef __OpenBSD__
-        for (int j = configData.tcpPorts[i].range.start; j <= configData.tcpPorts[i].range.end; j++) {
-          filter = ReallocAndAppend(filter, &filterLen, "tcp dst port %d", j);
-          if (j < configData.tcpPorts[i].range.end) {
-            filter = ReallocAndAppend(filter, &filterLen, " or ");
-          }
-        }
-#else
-        filter = ReallocAndAppend(filter, &filterLen, "tcp dst portrange %d-%d", configData.tcpPorts[i].range.start, configData.tcpPorts[i].range.end);
-#endif
-      }
-    }
-
-    if (configData.tcpPortsLength > 0 && configData.udpPortsLength > 0) {
-      filter = ReallocAndAppend(filter, &filterLen, ")");
-    }
-  }
-
-  if (configData.udpPortsLength > 0) {
-    if (configData.tcpPortsLength > 0 && configData.udpPortsLength > 0) {
-      filter = ReallocAndAppend(filter, &filterLen, " or (");
-    }
-
-    for (i = 0; i < configData.udpPortsLength; i++) {
-      if (i > 0) {
-        filter = ReallocAndAppend(filter, &filterLen, " or ");
-      }
-
-      if (IsPortSingle(&configData.udpPorts[i])) {
-        filter = ReallocAndAppend(filter, &filterLen, "udp dst port %d", configData.udpPorts[i].single);
-      } else {
-        /* OpenBSD's libpcap doesn't support portrange */
-#ifdef __OpenBSD__
-        for (int j = configData.udpPorts[i].range.start; j <= configData.udpPorts[i].range.end; j++) {
-          filter = ReallocAndAppend(filter, &filterLen, "udp dst port %d", j);
-          if (j < configData.udpPorts[i].range.end) {
-            filter = ReallocAndAppend(filter, &filterLen, " or ");
-          }
-        }
-#else
-        filter = ReallocAndAppend(filter, &filterLen, "udp dst portrange %d-%d", configData.udpPorts[i].range.start, configData.udpPorts[i].range.end);
-#endif
-      }
-    }
-
-    if (configData.tcpPortsLength > 0 && configData.udpPortsLength > 0) {
-      filter = ReallocAndAppend(filter, &filterLen, ")");
-    }
-  }
-
-  filter = ReallocAndAppend(filter, &filterLen, ")");
-
-  Debug("Device: %s pcap filter len %d: [%s]", device->name, filterLen, filter);
-
-  return filter;
-}
-
 int GetNoDevices(const struct ListenerModule *lm) {
   int count;
   struct Device *current;
@@ -360,8 +215,7 @@ void FreeListenerModule(struct ListenerModule *lm) {
 }
 
 int InitListenerModule(struct ListenerModule *lm) {
-  char errbuf[PCAP_ERRBUF_SIZE];
-  struct Device *current, *next;
+  struct Device *current;
 
   if (PrepDevices(lm) == FALSE) {
     return FALSE;
@@ -371,61 +225,8 @@ int InitListenerModule(struct ListenerModule *lm) {
 
   current = lm->root;
   while (current != NULL) {
-    next = current->next;
-
-    if ((current->handle = PcapOpenLiveImmediate(current->name, BUFSIZ, 0, BUFFER_TIMEOUT, errbuf)) == NULL) {
-      Error("Couldn't open device %s: %s", current->name, errbuf);
-      RemoveDevice(lm, current);
-      goto next;
-    }
-
-    if (pcap_setnonblock(current->handle, 1, errbuf) < 0) {
-      Error("Unable to set pcap_setnonblock on %s: %s", current->name, errbuf);
-      RemoveDevice(lm, current);
-      goto next;
-    }
-
-    /*
-     * OpenBSD and NetBSD has some quirks with pcap_setdirection(). Neither one of them will detect packets on the loopback interface
-     * if direction is set to PCAP_D_IN for example. There are some other inconsistencies as well and I might not have found all of them.
-     * By setting direction to PCAP_D_INOUT we make sure to capture as much as possible. The BPF filter will take care of most unwanted packets
-     * anyway so atleast for now, we set this to PCAP_D_INOUT on all platforms in order to avoid any potential missed packets.
-     */
-    if (pcap_setdirection(current->handle, PCAP_D_INOUT) < 0) {
-      Error("Couldn't set direction on %s: %s", current->name, pcap_geterr(current->handle));
-      RemoveDevice(lm, current);
-      goto next;
-    }
-
-    // We assume that since pcap_lookupnet() succeeded, we have a valid link type
-    if (pcap_datalink(current->handle) != DLT_EN10MB &&
-        pcap_datalink(current->handle) != DLT_RAW &&
-        pcap_datalink(current->handle) != DLT_NULL
-#ifdef __linux__
-        && pcap_datalink(current->handle) != DLT_LINUX_SLL
-#elif __OpenBSD__
-        && pcap_datalink(current->handle) != DLT_LOOP
-#endif
-    ) {
-      Error("Device %s is unsupported (linktype: %d), skipping this device", current->name, pcap_datalink(current->handle));
-      RemoveDevice(lm, current);
-      goto next;
-    }
-
-    if ((current->fd = pcap_get_selectable_fd(current->handle)) < 0) {
-      Error("Couldn't get file descriptor on device %s: %s", current->name, pcap_geterr(current->handle));
-      RemoveDevice(lm, current);
-      goto next;
-    }
-
-    if (SetupFilter(current) == FALSE) {
-      Error("Unable to setup filter for device %s, skipping", current->name);
-      RemoveDevice(lm, current);
-      goto next;
-    }
-
-  next:
-    current = next;
+    StartDevice(current);
+    current = current->next;
   }
 
   if (lm->root == NULL) {
@@ -617,47 +418,6 @@ struct Device *GetDeviceByFd(const struct ListenerModule *lm, const int fd) {
   }
 
   return NULL;
-}
-
-int SetupFilter(const struct Device *device) {
-  struct bpf_program fp;
-  char *filter = NULL;
-  int status = FALSE;
-  uint8_t isCompiled = FALSE;
-
-  assert(device != NULL);
-  assert(device->handle != NULL);
-
-  if ((filter = AllocAndBuildPcapFilter(device)) == NULL) {
-    goto exit;
-  }
-
-  // Using PCAP_NETMASK_UNKNOWN because we might use IPv6 and mask is only used for broadcast packets which we don't care about
-  if (pcap_compile(device->handle, &fp, filter, 1, PCAP_NETMASK_UNKNOWN) == PCAP_ERROR) {
-    Error("Unable to compile pcap filter %s: %s", filter, pcap_geterr(device->handle));
-    goto exit;
-  }
-
-  isCompiled = TRUE;
-
-  if (pcap_setfilter(device->handle, &fp) == PCAP_ERROR) {
-    Error("Unable to set filter %s: %s", filter, pcap_geterr(device->handle));
-    goto exit;
-  }
-
-  status = TRUE;
-
-exit:
-  if (filter != NULL) {
-    free(filter);
-    filter = NULL;
-  }
-
-  if (isCompiled) {
-    pcap_freecode(&fp);
-  }
-
-  return status;
 }
 
 static void PrintDevices(const struct ListenerModule *lm) {
